make initial move of files
