+++ mathesar/rpc/databases/privileges.py
@@ -3,7 +3,7 @@ from typing import Literal, TypedDict
 from modernrpc.core import rpc_method, REQUEST_KEY
 from modernrpc.auth.basic import http_basic_auth_login_required
 
-from db.roles.operations.select import list_db_priv, get_curr_role_db_priv
+from db.roles.operations.select import list_db_priv
 from db.roles.operations.update import replace_database_privileges_for_roles
 from mathesar.rpc.utils import connect
 from mathesar.models.base import Database
@@ -29,29 +29,7 @@ class DBPrivileges(TypedDict):
         )
 
 
-class CurrentDBPrivileges(TypedDict):
-    """
-    Information about database privileges for current user.
-
-    Attributes:
-        owner_oid: The `oid` of the owner of the database.
-        current_role_priv: A list of privileges available to the user.
-        current_role_owns: Whether the user is an owner of the database.
-    """
-    owner_oid: int
-    current_role_priv: list[Literal['CONNECT', 'CREATE', 'TEMPORARY']]
-    current_role_owns: bool
-
-    @classmethod
-    def from_dict(cls, d):
-        return cls(
-            owner_oid=d["owner_oid"],
-            current_role_priv=d["current_role_priv"],
-            current_role_owns=d["current_role_owns"]
-        )
-
-
-@rpc_method(name="database_privileges.list_direct")
+@rpc_method(name="databases.privileges.list_direct")
 @http_basic_auth_login_required
 @handle_rpc_exceptions
 def list_direct(*, database_id: int, **kwargs) -> list[DBPrivileges]:
@@ -71,27 +49,7 @@ def list_direct(*, database_id: int, **kwargs) -> list[DBPrivileges]:
     return [DBPrivileges.from_dict(i) for i in raw_db_priv]
 
 
-# TODO: Think of something concise for the endpoint name.
-@rpc_method(name="database_privileges.get_self")
-@http_basic_auth_login_required
-@handle_rpc_exceptions
-def get_self(*, database_id: int, **kwargs) -> CurrentDBPrivileges:
-    """
-    Get database privileges for the current user.
-
-    Args:
-        database_id: The Django id of the database.
-
-    Returns:
-        A dict describing current user's database privilege.
-    """
-    user = kwargs.get(REQUEST_KEY).user
-    with connect(database_id, user) as conn:
-        curr_role_db_priv = get_curr_role_db_priv(conn)
-    return CurrentDBPrivileges.from_dict(curr_role_db_priv)
-
-
-@rpc_method(name="database_privileges.replace_for_roles")
+@rpc_method(name="databases.privileges.replace_for_roles")
 @http_basic_auth_login_required
 @handle_rpc_exceptions
 def replace_for_roles(
