+++ mathesar/rpc/roles/configured.py
@@ -29,7 +29,7 @@ class ConfiguredRoleInfo(TypedDict):
         )
 
 
-@rpc_method(name="configured_roles.list")
+@rpc_method(name="roles.configured.list")
 @http_basic_auth_login_required
 @handle_rpc_exceptions
 def list_(*, server_id: int, **kwargs) -> list[ConfiguredRoleInfo]:
@@ -47,7 +47,7 @@ def list_(*, server_id: int, **kwargs) -> list[ConfiguredRoleInfo]:
     return [ConfiguredRoleInfo.from_model(db_model) for db_model in configured_role_qs]
 
 
-@rpc_method(name='configured_roles.add')
+@rpc_method(name='roles.configured.add')
 @http_basic_auth_superuser_required
 @handle_rpc_exceptions
 def add(
@@ -77,7 +77,7 @@ def add(
     return ConfiguredRoleInfo.from_model(configured_role)
 
 
-@rpc_method(name='configured_roles.delete')
+@rpc_method(name='roles.configured.delete')
 @http_basic_auth_superuser_required
 @handle_rpc_exceptions
 def delete(*, configured_role_id: int, **kwargs):
@@ -91,7 +91,7 @@ def delete(*, configured_role_id: int, **kwargs):
     configured_role.delete()
 
 
-@rpc_method(name='configured_roles.set_password')
+@rpc_method(name='roles.configured.set_password')
 @http_basic_auth_superuser_required
 @handle_rpc_exceptions
 def set_password(
